+++ packages/app/package.json
@@ -27,9 +27,7 @@
   "dependencies": {
     "@pixi/core": "^5.0.0-alpha.2",
     "@pixi/display": "^5.0.0-alpha.2",
-    "@pixi/settings": "^5.0.0-alpha.2",
-    "@pixi/ticker": "^5.0.0-alpha.2",
-    "@pixi/interaction": "^5.0.0-alpha.2"
+    "@pixi/settings": "^5.0.0-alpha.2"
   },
   "devDependencies": {
     "@pixi/canvas-renderer": "^5.0.0-alpha.2",
