+++ packages/app/src/ResizePlugin.js
