+++ packages/app/src/index.js
@@ -1,10 +1,6 @@
 import Application from './Application';
-import TickerPlugin from './plugins/TickerPlugin';
-import ResizePlugin from './plugins/ResizePlugin';
-import InteractionPlugin from './plugins/InteractionPlugin';
+import ResizePlugin from './ResizePlugin';
 
-Application.registerPlugin(TickerPlugin);
 Application.registerPlugin(ResizePlugin);
-Application.registerPlugin(InteractionPlugin);
 
 export { Application };
