+++ packages/interaction/src/InteractionPlugin.js
