+++ packages/ticker/src/TickerPlugin.js
